pyproject.toml
[tool.poetry]
name = "langchain-practice"
version = "0.1.0"
description = ""
authors = ["youngchan-itall <michael.yc.cho@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
langchain = "^0.3.9"
langchain-openapi = "^0.1.1"
langchain-community = "^0.3.9"
langchainhub = "^0.1.21"
black = "^24.10.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
